+++ KarteUtilities/Networking/HTTPMethod.swift
@@ -1,5 +1,5 @@
 //
-//  Copyright 2020 PLAID, Inc.
+//  Copyright 2025 PLAID, Inc.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -13,26 +13,11 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 //
+//  NOTE: The Implementation was inspired by APIKit.
 
 import Foundation
-import KarteUtilities
 
-internal protocol TrackClientSession {
-    @discardableResult
-    func send(
-        _ request: TrackRequest,
-        handler: @escaping (Result<TrackRequest.Response, NetworkingError>) -> Void
-    ) -> URLSessionTask?
-}
-
-internal class DefaultTrackClientSession: TrackClientSession {
-    func send(
-        _ request: TrackRequest,
-        handler: @escaping (Result<TrackRequest.Response, NetworkingError>) -> Void
-    ) -> URLSessionTask? {
-        return Session.send(request, handler: handler)
-    }
-
-    deinit {
-    }
+public enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
 }