+++ CHANGELOG.md
@@ -2,16 +2,30 @@
 
 | モジュール名 | Description | 最新のバージョン |
 | :-- | :-- | :-- |
-| KarteCore | イベントトラッキング機能を提供します。 | 2.30.0 |
+| KarteCore | イベントトラッキング機能を提供します。 | 2.31.0 |
 | KarteInAppMessaging | アプリ内メッセージ機能を提供します。 | 2.19.0 |
 | KarteRemoteNotification | プッシュ通知の受信および効果測定機能を提供します。 | 2.11.0 |
 | KarteVariables | 設定値配信機能を提供します。 | 2.11.0 |
-| KarteVisualTracking | ビジュアルトラッキング機能を提供します。 | 2.12.0 |
+| KarteVisualTracking | ビジュアルトラッキング機能を提供します。 | 2.13.0 |
 | KarteInAppFrame | アプリ内埋め込みコンポーネント機能を提供します。 | 0.4.0 |
 | KarteCrashReporting | クラッシュイベントのトラッキング機能を提供します。 | 2.9.0 |
-| KarteUtilities | KarteCore モジュール等が利用するUtility機能を提供します。通常直接参照する必要はありません。 | 3.12.0 |
+| KarteUtilities | KarteCore モジュール等が利用するUtility機能を提供します。通常直接参照する必要はありません。 | 3.13.0 |
 | KarteNotificationServiceExtension | リッチプッシュ通知機能を提供します。 | 1.2.0 |
 | KarteDebugger | KARTE for Appのデバッグ機能を提供します。 | 1.0.0 |
+
+# Releases - xxxx.xx.xx
+
+## Utilities 3.13.0
+** 🔨CHANGED**
+- 一部のAPI通信の実装を3rd partyライブラリから内製のコードに置き換えました。
+
+### Core 2.31.0
+** 🔨CHANGED**
+- API通信実装変更に伴う変更を行いました。
+
+## Visual Tracking 2.13.0
+** 🔨CHANGED**
+- API通信実装変更に伴う変更を行いました。
 
 # Releases - 2025.05.26
 ## Version 2.24.0